+++ synapse/lib/stormlib/macro.py
@@ -125,6 +125,30 @@ class LibMacro(s_stormtypes.Lib):
     '''
     A Storm Library for interacting with the Storm Macros in the Cortex.
     '''
+    _storm_locals = (
+        {'name': 'set', 'desc': 'Add or modify an existing Storm Macro in the Cortex.',
+         'type': {'type': 'function', '_funcname': '_funcMacroSet',
+                  'args': (
+                      {'name': 'name', 'type': 'str', 'desc': 'Name of the Storm Macro to add or modify.', },
+                      {'name': 'storm', 'type': ['str', 'storm:query'], 'desc': 'The Storm query to add to the macro.', },
+                  ),
+                  'returns': {'type': 'null', }}},
+        {'name': 'get', 'desc': 'Get a Storm Macro definition by name from the Cortex.',
+         'type': {'type': 'function', '_funcname': '_funcMacroGet',
+                  'args': (
+                      {'name': 'name', 'type': 'str', 'desc': 'The name of the macro to get.', },
+                  ),
+                  'returns': {'type': 'dict', 'desc': 'A macro definition.', }}},
+        {'name': 'del', 'desc': 'Delete a Storm Macro by name from the Cortex.',
+         'type': {'type': 'function', '_funcname': '_funcMacroDel',
+                  'args': (
+                      {'name': 'name', 'type': 'str', 'desc': 'The name of the macro to delete.', },
+                  ),
+                  'returns': {'type': 'null', }}},
+        {'name': 'list', 'desc': 'Get a list of Storm Macros in the Cortex.',
+         'type': {'type': 'function', '_funcname': '_funcMacroList',
+                  'returns': {'type': 'list', 'desc': 'A list of ``dict`` objects containing Macro definitions.', }}},
+    )
     _storm_lib_path = ('macro',)
 
     def getObjLocals(self):
@@ -136,40 +160,16 @@ class LibMacro(s_stormtypes.Lib):
         }
 
     async def _funcMacroList(self):
-        '''
-        Get a list of Storm Macros in the Cortex.
-
-        Returns:
-            list: a list of dictionaries representing Macro definitions.
-        '''
         path = ('cortex', 'storm', 'macros')
         return await self.runt.snap.core.getHiveKeys(path)
 
     async def _funcMacroGet(self, name):
-        '''
-        Get a Storm Macro definition by name from the Cortex.
-
-        Args:
-            name (str): The name of the macro to get.
-
-        Returns:
-            dict: A macro definition.
-        '''
         name = await s_stormtypes.tostr(name)
 
         path = ('cortex', 'storm', 'macros', name)
         return await self.runt.snap.core.getHiveKey(path)
 
     async def _funcMacroDel(self, name):
-        '''
-        Delete a Storm Macro by name from the Cortex.
-
-        Args:
-            name (str): The name of the macro to delete.
-
-        Returns:
-            dict: The macro definition which has been removed from the Cortex.
-        '''
         name = await s_stormtypes.tostr(name)
         path = ('cortex', 'storm', 'macros', name)
 
@@ -186,17 +186,6 @@ class LibMacro(s_stormtypes.Lib):
         await self.runt.snap.core.popHiveKey(path)
 
     async def _funcMacroSet(self, name, storm):
-        '''
-        Add or modify an existing Storm Macro in the Cortex.
-
-        Args:
-            name (str): Name of the Storm Macro to add or modify.
-
-            storm (str): The Storm query to add to the macro.
-
-        Returns:
-            None: Returns None.
-        '''
         name = await s_stormtypes.tostr(name)
         storm = await s_stormtypes.tostr(storm)